# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Films API",
        "version": "0.1"
    },
    "paths": {
        "/api/critics/{id}": {
            "delete": {
                "tags": [
                    "Critics"
                ],
                "summary": "Delete existing Critic",
                "operationId": "4c7bbd907e57760bbbfc560c1ba8e1ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of critic to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/films/{id}/actors": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Gets a movie's actors",
                "operationId": "1767c0ecf30e192d1da1267965ce4864",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/films": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Gets list of movies",
                "operationId": "271e474415d2f29bf3c163247c5f4639",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword in wanted movie's name"
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Rating of the wanted movie"
                    },
                    {
                        "name": "minLength",
                        "in": "query",
                        "description": "Minimum length of the wanted movie"
                    },
                    {
                        "name": "maxLength",
                        "in": "query",
                        "description": "Maximum length of the wanted movie"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/films/{id}/average_score": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Gets a movie's average score",
                "operationId": "68193bce2626a6718b4afcdfe4f8e66c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/films/{id}/critics": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Gets a movie's critics",
                "operationId": "e456bf562a66ec61222dab0ca5c0a3c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Creates a user",
                "operationId": "1a63883edb751cb640a2f6f516236f49",
                "requestBody": {
                    "content": {
                        "application\\json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Updates a user",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\\json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/users/{id}/favorite_language": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Gets a user's favorite language",
                "operationId": "d41b63b6f7d503c0b0a8a0ffc5e8144b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "404": {
                "description": "Not found"
            }
        }
    }
}